refactored code and separated shoppinglist component to additem and listview components
